Add files via upload
This new update officially makes the jump between the two API feeds and connects two games to ensure they are the same game.  The only issue now double headers.  So I"ll need to add a few more logical statements to check for double headers.  There is a field in the MLB api that checks for multiple games in in a day.